requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
opensearch-py>=2.6.0
chromadb>=0.5.5
cachetools>=5.3.3
tiktoken>=0.7.0
pypdf>=4.3.1
python-docx>=1.1.2
beautifulsoup4>=4.12.3
pandas>=2.2.2
streamlit>=1.36.0
opentelemetry-api>=1.26.0
opentelemetry-sdk>=1.26.0
opentelemetry-exporter-otlp>=1.26.0
odfpy>=1.4.1
ragas>=0.2.0
datasets>=2.18.0
langchain-openai>=0.2.0
sentence-transformers>=2.2.2
